refactor: move dot logic out of main mod
